Person class implements Comparable interface
Example of use of java.util.PriorityQue class with preservation of insertion order.

Let's forget about how Java implements a PriorityQueue and treat it as a black box.  We know that it does not preserve insertion order. For example, create a PriotityQueue of Integers and add 5, 7, 3 to it in that order. When you remove the elements one-by-one, you get back 3, 5, 7 (in ascending order).  Similarly, a PriorityQueue of strings filled with "b", "a", "c" will return "a", "b", "c."  Thus numbers and strings are compared by virtue of their "natural" sort order.

On the hand, if we define a custom class and store its instances in a PriorityQue, we have to define how Java compares these objects. We do so by having that class implement the Comparable interface and overriding the compareTo method in that class. Consider a class Person that only has String name and int age as its instance variables. We construct Person objects joe 30, mary 20, and tom 40 and store them into a PriorityQue of Person objects.  If we want the persons PQ to use age as the priority and return objects in ascending order of age, we have to have: class Person implements Comparable<Person>. In the body of its compareTo method we simply put return this.age - other.age. This method returns an integer that, based on its negative or positive sign, determines the order of this vs. other. After doing this, the outcome of repeated removals from the priority queue will be mary 20, joe 30, tom 40.
  
Now, onto the challenge. Let's say after adding those 3 people to the persons priority queue, we also add jane 30 and bob 30. Now we have three people (joe, jane, bob) that have the same priority.  When removed, Java arbitrarily determines how to remove objects that are tied in priority. It may return mary 20, bob 30, joe 30, jane 30, tom 40 or some other permutation of joe, jane and bob and ignores insertion order. How do we enforce insertion order? Let's say we want to let younger people into a club before older people but, if we have two or more people of the same age, we want to admit them in the order they got in line.
  
The solution is to declare another instance variable in the Person class called insertionPriority and set that variable at the moment of instantiation/insertion into the queue. We would also have to have a static variable called insertionOrder that is incremented in the constructor and keeps track of how many objects have been instantiated already. Finally, we write the compareTo method with a simple if/else conditional such that, if the ages are equal, it compares this and other based on their insertionPriority. I have put the code for this here.